+++ src/formatters/html/TylerHtmlCompressor.php
@@ -18,9 +18,37 @@ use yii\base\Component;
 class TylerHtmlCompressor extends Component implements IFormatter
 {
 
+    /**
+     * Perform extra (possibly unsafe) compression operations
+     * @var bool
+     */
+    public $extra = false;
+
+    /**
+     * Removes HTML comments
+     * @var bool
+     */
+    public $noComments = true;
+
+    /**
+     * @param string $html
+     * @return string
+     */
     public function format($html)
     {
-        return HtmlCompressor::compress($html, []);
+        $options = [
+            'no-comments' => $this->noComments,
+            'extra' => $this->extra,
+        ];
+
+        \Yii::beginProfile('countHtmlStrings');
+            $count = substr_count($html, "\n") + 1;
+            \Yii::info('Number of HTML strings: ' . $count);
+        \Yii::endProfile('countHtmlStrings');
+
+        $result = HtmlCompressor::compress((string) $html, $options);
+
+        return $result;
     }
 
 }